pyproject.toml
[project]
name = "product"
version = "0.1.0"
description = ""
authors = [
    {name = "www",email = "pyostr@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "infisicalsdk (>=1.0.16,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic-settings (>=2.13.0,<3.0.0)",
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "gunicorn (>=25.1.0,<26.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "cryptography (>=46.0.5,<47.0.0)",
    "boto3 (>=1.42.53,<2.0.0)",
    "python-multipart (>=0.0.20,<0.1.0)",
    "redis (>=6.4.0,<7.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "numpy (>=2.4.2,<3.0.0)",
    "opencv-python (>=4.13.0.92,<5.0.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "tqdm (>=4.67.3,<5.0.0)",
    "rembg[cpu] (>=2.0.72,<3.0.0)",
]

[tool.poetry.group.dev.dependencies]
pytest = ">=9.0.2,<10.0.0"
pytest-asyncio = ">=1.3.0,<2.0.0"
isort = ">=7.0.0,<8.0.0"
testcontainers = {version=">=4.14.1,<5.0.0", extras=["postgresql"]}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
